Add: made /data directory and move OddOut data file
